Refactor views folder
