+++ js/viz/views/positionItems/positionAssociation.js
@@ -1,28 +1,5 @@
-import { schemaWithLinks } from "../../../state/globals.js";
-
-export function layoutMonteCarloReco(viewObjects) {
-  const typeName = schemaWithLinks()
-    ? "podio::LinkCollection<edm4hep::ReconstructedParticle,edm4hep::MCParticle>"
-    : "edm4hep::MCRecoParticleAssociation";
-  return layoutAssociation(viewObjects, typeName);
-}
-
-export function layoutMonteCarloTrack(viewObjects) {
-  return layoutAssociation(
-    viewObjects,
-    "edm4hep::MCRecoTrackParticleAssociation",
-  );
-}
-
-export function layoutMonteCarloCluster(viewObjects) {
-  return layoutAssociation(
-    viewObjects,
-    "edm4hep::MCRecoClusterParticleAssociation",
-  );
-}
-
 // List 1:1 association in a vertical list
-function layoutAssociation(viewObjects, associationName) {
+export function positionAssociation(viewObjects, associationName) {
   const associations = viewObjects.associations[associationName];
 
   const fromWidth = associations[0].from.width;
